requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.6.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
decorator==5.1.1
Django==4.1.7
django-discover-runner==1.0
django-filter==23.1
django-role-permissions==3.1.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.4.0
protobuf==4.22.1
psycopg2==2.9.6
psycopg2-binary==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
Pyrogram==2.0.102
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==13.0
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
TgCrypto==1.2.5
tornado==6.2
tzdata==2023.2
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
